Fix redux "Store does not have a valid reducer" error (#80)
Related to #71 . When the registry was added, it was first creating the store with no reducers, which caused this error to be printed out. Now, add reducers to the registry before creating the store to avoid the error.

Note reducers can still be added to the registry after that fact.