WIP Add a MatPlotLib plugin to render simple image data from matplotlib
This is pretty disgusting but it works as a proof of concept

WIP Try and handle custom objects right. It's not right yet

Scale the image to fill the panel

Now you don't need to scroll around or whatever.

At some point will want to pass to the backend how big of a panel we have so it can resize accordingly