DH-11788 Fix aggregation formatting when only one aggregation is applied (#580)
When there is only one aggregation operation applied, it does not append the operation as a suffix to the column name in the totals table. Fix it up so it looks for the correct column name when there is only one operation.

This was noticed with the Count operation since the totals column format doesn't always match the source column format with Count, but happened with any type of operation.