feat: adds a custom linker-not-allowed cursor (#1060)
Browser not-allowed didn't visually pair well with the linker cursor it
switches between when moving across columns. Adds a cursor that combines
a not allowed cursor with a cross hair of the same style of our linker
cursor. Icon is now also BW only, as the danger color was a stronger
warning than needed.