feat: Instants and ZonedDateTimes should be treated as DateTimes (#1117)
deephaven/deephaven-core#3385 is a port of DH-11692 which adds `Instant`
and `ZonedDateTime` support to the engine. I've similarly encoded these
types over Barrage as a `long` (as a nanosecond since epoch). The
encoding of Zones will actually need to be implemented at a future date
as the best implementation requires (encourages?) a much larger
refactoring of ColumnSource types. See deephaven/deephaven-core#3455 for
more information.

I was able to get the web-client-ui to display these types with this
small patch.