feat: Improve plugin load error handling (#1214)
Loads plugins individually instead of failing to register if any plugin
fails to load. Previously if any plugin failed to load for any reason
then no plugins were registered.

Does some validation on if the manifest file exists and the structure of
it so that the errors are more descriptive. Does not throw a 2nd error
about `Unexpected token <` if there is no manifest file