build: Consume built packages in apps instead of bundling from src (#1339)
This addresses our sprint retro item of a test to detect packaging
issues.

By changing the apps to not alias to src files in build mode, the DHC
code-studio should see packages the same way the enterprise web-ui build
will. This should help detect most packaging issues since the majority
of the code in DHC packages are consumed by code-studio.

This seems to make the bundle slightly larger. Probably Vite was able to
de-dupe more when bundling all of the packages from src

As a test, cherry-pick
https://github.com/deephaven/web-client-ui/commit/d1838d7151e31de1e26da8b60f336e991bb25404
onto this branch and run build. Should get the same error mentioned in
#1334 when you run build, but not when you run dev mode