test: Update e2e docker template (#1427)
While helping Ethan debug e2e flaky tests in CI, Don mentioned the npm
script to run e2e in docker was ugly. I agreed and decided to clean it
up some.

This change makes the e2e tests in Docker also spin up the DHC server w/
anonymous auth so you don't have to. We could extend this to add a dev
mode for e2e testing possibly, but I don't think you can get headed mode
in Docker if you want it.

You can also pass arguments directly now. `npm run e2e:ci --
--project=chromium tests/someTest.js` and it will run in Docker

Also figured out a way to remove a COPY line for every package.json
using multi-stage builds. It caches properly still if no package.json
files change.

Removed the volume mount for test-results because Docker creates the
directories inside of it as root (I tried running the container as a
regular user and it complains). This causes an issue if you run the
snapshots in Docker w/ a failure and then try to run locally because the
folder is owned by root and can't be cleared as a regular user.

Still need to update the readme. And probably rename `e2e:ci` since it's
not actually run in CI