fix: Remove unused ref forwarded to all dashboard panels (#1451)
This ref doesn't seem to be used by anything and prints dev warnings if
any panel is a functional component that doesn't use `React.forwardRef`.
I couldn't find anywhere we actually use the ref and it looks like it
was a workaround for golden-layout adding the ref as default props to
every component. The way we use golden-layout, I don't think it's even
possible to attach a ref to the component in a way that a plugin could
access that ref.