fix: Zip CSV uploads not working (#1457)
Fixes #1080. Fixes #1416 

Updates jszip and uses the internal stream helper instead of nodestream
which we would need a polyfill for. The progress on zip uploads is a bit
odd since we use the unzip progress. It seems papaparse loads more from
the stream while processing/before processing and can finish reading the
zip before it's done processing chunks.

Also, uploading the tables seems to be a blocking operation. Not sure if
that's an easy fix, but after the 50% mark on uploading a large zip,
there are some blocks on the main thread (marching ants freezes as an
indicator).

There is a ~2GB limit for JSZip it seems
https://github.com/Stuk/jszip/issues/777