fix: Plugin peer dependencies do not get versions from lerna (#1517)
Apparently lerna doesn't treat `file` items in peer dependencies the
same as it does for regular or dev dependencies. They don't get replaced
and make the peer dependency unfindable.

Compare these 2. In peer dependencies for plugin it lists
`file:../dashboard` instead of an actual version. But for regular
dependencies in components, they get a version.

https://unpkg.com/@deephaven/plugin@0.48.0/package.json
https://unpkg.com/@deephaven/components@0.48.0/package.json