fix: Theming - switched from ?inline to ?raw css imports (#1600)
It seems that Enterprise does not handle css ?inline css imports in
node_modules the same way it does in src. Namely, the raw content gets
imported as an empty string instead of the raw content.

- Converted ?inline imports to ?raw. These seem to handle the raw
content as expected even if in node_modules
- Added a sass minification step to @deephaven/components. Vite seems to
minify ?inline imports but does not do so with raw
- Added a try / catch to MonacoUtils to avoid non-hex theme variables
from crashing the entire app

fixes #1599

BREAKING CHANGE: Theme css imports were switched from `?inline` to
`?raw`. Not likely that we have any consumers yet, but this would impact
webpack config.