fix: non-contiguous table row selection background colour (#1623)
Fixes #1619

Draw selection was calling the fill/stroke inside the loop, which
implicitly calls closePath, resulting the fill to be drawed over top of
itself multiple times. Move the fill/stroke outside the loop the creates
the selection rects.

Adds an e2e test that makes a non-contiguous selection to prevent
regressions.