feat: theme fontawesome icon size wrapped in spectrum icons
Spectrum icons are 18px. However, our icons look best at 16px as they were designed on a 16px grid. This overrides the default spectrum icon size when wrapping a font-awesome icon inside a spectrum icon wrapper.

Made overrides not a module, so it doesn't add hashes to the classes in overrides, as I have a feeling this file will be used to target specific classes and that wouldn't be desirable.

Note: please suggest somewhere for that variable I added to live if you think that location sucks