fix: popper blur in styleguide
Poppers such as menu dropdowns aren't closing in the styleguide on outside clicks, this is because no element has focus on the outside click, as e.relatedTarget was null. The main app doens't have this problem since tabindex=-1 is on the appmaincontainer. Add a wrapper (that includes the side marigins) to have a -1  tabindex.