test: fix flaky tests (#1803)
### Context Menu

- Fixes #1799 
- Add a check that `.loading-spinner` disappears before locating and
clicking the Go To button

Multiselect values (#1736) turned the filter options to be async. This
results in the context menu having a loading spinner at the bottom for a
few milliseconds. Flakiness happens when the filter options are resolved
between the Go To value being located and pressed.


![image](https://github.com/deephaven/web-client-ui/assets/93849557/54af9484-e794-4bf1-b098-d65d833ca2d4)

![image](https://github.com/deephaven/web-client-ui/assets/93849557/3e192f86-4142-4674-8920-cd412d5dffa1)


### Multiselect Filter Flakiness

- Add a check for the loading spinner after opening a table/plot
- Add a check for .`iris-grid-loading-status-bar` after applying filter

Upon opening a table/plot, the loading indicator is a `.loading-spinner`
instead of `.iris-grid-loading-status`. As such, the check is passed.
This causes the mouse click to happen even though the grid isn't loaded
yet and Playwright is waiting for a context menu to appear when
`.iris-grid-loading-status` is present.


![image](https://github.com/deephaven/web-client-ui/assets/93849557/b05ab37d-3463-47d0-b9b2-0c5e16045336)

![image](https://github.com/deephaven/web-client-ui/assets/93849557/81c6c9d9-0788-4450-aadd-f6a7583387f0)