chore: Add buffer as a chart dependency (#1899)
This adds an explicit dependency to `buffer` to `@deephaven/chart` which
we currently get implicitly from `browserify` in `golden-layout`.
`plotly.js` depends on `typedarray-pool` which depends on `buffer`, but
doesn't list it.

I opened a PR with `typedarray-pool`, but doesn't seem it's an active
repo any more. https://github.com/mikolalysenko/typedarray-pool/pull/11

The docs uses webpack 5 w/ docusaurus and I added `buffer` when adding
chart support, so this should be all that's needed.
https://github.com/deephaven/deephaven.io/commit/6a8bcf29fe32dea1825c6d7f682857cc33b59109#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519