feat: Embed widget loading workspace settings (#2068)
Fixes #1964. Also removed a bunch of unnecessary code in `AppInit` where
we were using the redux `connect` API with a functional component
instead of just using the redux hooks

Note that even though this moves a bunch of files to `app-utils`, I
don't consider it breaking since they are being moved out of
`code-studio`. They wouldn't be consumed by anything else since
`code-studio` is published as a bundle.

Due to local storage requiring a port match, this can only be tested w/
the local server in 2 steps (this should be fine in prod since the app
and embed-widget run on the same port there).

1. Start the dev server and go to the app
2. Run some code to create a table.
3. Change user formatting settings (e.g. show timezones and have a
timestamp column)
4. Stop the dev server.
5. Run `PORT=4000 npm run start:embed-widget`. This will let us access
the saved workspace from the app
6. Go to `localhost:4000/?name=<your-variable-name>`
7. The formatting settings should be applied