refactor: Make IrisGridProxyModel use an actual Proxy (#2085)
This makes `IrisGridProxyModel` use an actual Proxy instead of requiring
us pass through most of the calls to the underlying model.

The proxy will check if the `IrisGridProxyModel` class implements a
method first. If it does, then it calls the implemented method. If it
does not, it delegates the call to the current `model`

Many of the methods which had a check for a specific model type first
can also be handled by the proxy. A caller is must use that check first
anyway or else TS will complain. So the check was really being done
twice before

Didn't look into it too hard, but I disabled the TS check for `extends
IrisGridModel`. There are many abstract methods in `IrisGridModel` which
are implemented by the proxy handling them, but TS doesn't know that.