fix: GridMetrics including totals rows in visible rows when scrolled to bottom (#2194)
Noticed this while working on databars. The totals rows are being
included in `metrics.visibleRows` when scrolled to the bottom. They were
also being duplicated in `allRows`.

1. Create a table such as `t = empty_table(1000).update(["X=ii"])`
2. Add an aggregation
3. Scroll to bottom of the table
4. Log `visibleRows` or `allRows` from `GridMetrics`

Without this change you'll see that the totals row(s) (index 1000+) are
included in those arrays when scrolled to the bottom. They are doubled
in `allRows`

This doesn't seem to be the issue when aggregations are set to the top
and the table is scrolled to the top.