refactor: Remove @adobe/react-spectrum imports where prohibited (#2198)
Fixes #2195 

Had to just ignore some of the errors because we compare spectrum
components in the styleguide and I didn't want to mess with that for
now.

I still think it's a bit confusing that there are things in
`components/src/spectrum` that aren't pure re-exports (e.g. ComboBox).
Makes it confusing to know which to use.