feat: Deephaven UI table databar support (#2190)
Changes needed for dh.ui databars

- Support for extracting data in viewport columns not listed in
`table.columns`. This allows us to hide the aggregation columns for
databars from the user but still get their data. If a column is not in
`table.columns` then it will be added to viewport data via its name
instead of model index
- Improved gradient rendering to be much more efficient
- Fixed opacity being ignored if a gradient was used
- Fixed some of the databar spacing (was 1px off center)
- Modified `colorValueStyle` to have overrides so if we pass a string we
don't get back `string | undefined` when we know we'll get a string
- Modified `resolveCssVariablesInRecord` to always resolve CSS colors
regardless of if they contain a variable