feat: Ruff Configuration Settings (#2215)
Fixes #1255 

I need to test this doesn't crash if using groovy. Not sure if the ruff
version is available without initializing ruff first or what happens if
it's unavailable.

- Added global setting for minimap, format on save, formatter, and
formatting settings
- Added format on save option and format button to notebook overflow
menu
- Format on save triggers on both shortcut and pressing the save button
- Config editor has JSON schema included
- Config editor cannot be closed by clicking outside the modal, but can
be closed w/ escape key

Something we should add maybe as part of this PR or a final follow-up to
this feature branch is disabling certain/all linting/formatting for the
console. One way we might do this is by registering the console w/ a
specific URI or URI scheme so we can check it when checking if we should
lint. Or include it as a prop to the `MonacoProviders` component. This
would require some changes to the current setup of setting/using the
config on `MonacoProviders` static members/methods

(This might not actually be breaking, it's more just notes for
implementing in DHE)
BREAKING CHANGE:
The app should call `MonacoUtils.init` with a `getWorker` function that
uses the JSON worker in addition to the general fallback worker when
adding support for configuring ruff.