fix: Layout shifts when opening and closing panels from a fresh state (#2241)
Fixes #1268

This mostly occurs when you have a relatively fresh state. There's a
video in the ticket, but if you have just 2 rows then this is frequently
triggered. Using incognito mode should give you a fresh state which is
affected by this bug.

This fixes both the vertical shift when you open a new panel type (open
a notebook from a fresh state or open a table from a fresh layout state)
and when you close the last panel in a stack (like in the ticket)