fix: Change ruff errors to warnings and fix config saving (#2246)
Changed all except syntax errors to warning underlines instead of error
underline since they are not actual errors.

Also changed redux to not save if the config matches the default. This
should fix an edge case where a user opens the settings editor just to
see what the config is, then hits "Save". Previously, this would save
the current defaults as their settings and then if the defaults changed
they would not get updated settings. This way, if a user was on defaults
and the default changes they will get the new default.

Also added a little bit of debug logging for Ruff.