feat: Add render blocking errors to Chart (#2255)
Adds new events that enable render-blocking overlays to the chart.
The current error pop-up is not sufficient because it still allows the
chart to render. This is problematic in the case of webgl because
disabling webgl means we don't want the chart rendering at all in case
the computer can't handle it.