refactor: Change iris-grid reverseType to reverse and deprecate reverseType (#2278)
Fixes #2149 

This shouldn't be breaking. `reverseType` prop is deprecated, but still
accepted. `reverseType` in saved states will be changed into `reverse`
and then dropped from the saved state.

Tested this by exporting a layout with a table with reverse from the
current DHC release. Then imported the layout to my dev UI and checked
the reverse persisted.