fix: DH-18542: Remove duplicate and invalid ruff quick fixes (#2360)
Should remove the invalid quick fixes to disable syntax errors (wouldn't
that be nice if you could do that)

Also if there is a duplicate code detected (usually only if you have a
selection and use the light bulb instead of the hover quick fix link in
the modal), then the "diasble RULE for file" should only appear once.
There should also be "disable RULE for line X" and "disable RULE for
line Y" instead of just "for line".