refactor: Require memoizee and memoizeClear calls have a max size (#2375)
Related to #2374 (and might need that change before this can merge).
This requires all `memoizee` and `memoizeClear` calls to provide a max
cache size. Most of these were small caches most likely, but this should
help prevent another issue that is hard to spot since it takes
potentially hours to reproduce like DH-18798