feat: add median agg for totals table (#2388)
https://deephaven.atlassian.net/browse/DH-18856

Requires this change to jsapi (tested):
https://github.com/deephaven/deephaven-core/pull/6700

Median works with Totals Table, but does not work with rollups (not
supported in the engine). After discussion with Don, the agg is ignored
for rollups and displays a warning to such an effect. A lot of the logic
changes exist to support that behavior.