refactor: DH-19267: columns changed filter event takes an id rather than a panel (#2417)
https://deephaven.atlassian.net/browse/DH-19267

Refactors `InputFilterEvent.COLUMNS_CHANGED` to take an id instead of a
panel. This can be a panel id or a widget id. This will allow
`deephaven.ui` components to send this event. Note that more work will
need to be done on Linker. This will be done later when we address
https://deephaven.atlassian.net/browse/DH-18840