feat: Add usePersistentState hook (#2402)
Adds the `usePersistentState` hook and `PersistentStateProvider` which
used together track calls to `usePersistentState` and calls an
`onChange` when the persisted state changes. The state is persisted as
an array of the calls to the hook within the provider.

If the tree initially renders in a different order or with fewer/more
hook calls, then the rendering may break depending on what state is now
swapped. We could easily detect if the number of calls to
`usePersistentState` changed on a hydration and throw an error or
re-render the component from scratch, but right now the implementation
does not do that.

This should handle memoized components, nested components below the
provider, and components which immediately re-render via a conditional
setState (`useEffect` in the provider is only fired once even though the
component renders twice).

It is technically possible to use this with class components via context
and mimicking the logic in the hook by calling `addState` in `render`
and `trigger` anywhere that should trigger a state change, but I think
it would be best to just wrap the few class components that need this in
a thin functional wrapper to use the hook.

If there is no provider, `usePersistentState` still functions
effectively just as `useState`. This means we should be able to call the
hook in widgets that can be used standalone and in dh.ui without issue
(like plotly express or IrisGrid)