fix: DH-19350: Ruff tries to lint before initializing when disabled (#2429)
Looks like it was an issue caused if Ruff was disabled when we changed
settings (we set settings on initial load). This was causing a check to
pass and then trying to use Ruff without initializing it

Also added some log suppression because Ruff internals call
`console.log` when the workspace is created and it's not useful to us.
It's something along the lines of `glob 25 created ...` and I just found
it unnecessary noise