feat: Add keepMounted param to TabPanels for persisting panels when they are not active (#2434)
Part of DH-18349. Will need a small update in dh.ui to enable support
there which is what the ticket wants.

Does not change default behavior of `TabPanels`, just adds an extra
param to keep static panels mounted (in React, but not actually mounted
in the DOM).

There are unit tests, but can also test with local plugins from this PR
https://github.com/deephaven/deephaven-plugins/pull/1177