refactor: Change iris-grid reverseType to reverse and deprecate reverseType (#2278) (#2447)
Cherry pick 3 of 5 (I think) for usePersistentState. Wasn't a completely
clean pick, but verified against the original PR that the expected
changes are all that changed.

This shouldn't be breaking. `reverseType` prop is deprecated, but still
accepted. `reverseType` in saved states will be changed into `reverse`
and then dropped from the saved state.

Tested this by exporting a layout with a table with reverse from the
current DHC release. Then imported the layout to my dev UI and checked
the reverse persisted.