feat: Add IrisGridCacheUtils for memoizing iris grid state (#2416) (#2448)
Cherry pick 4 of 5 for usePersistentState

Adds `IrisGridCacheUtils` which provides 3 methods to create memoized
dehydrators: `makeMemoizedGridStateDehydrator`,
`makeMemoizedIrisGridStateDehydrator`, and
`makeMemoizedCombinedGridStateDehydrator`. The first 2 are used for
`IrisGridPanel` which saves the state under separate keys for
`gridState` and `irisGridState`. The 3rd is for newer uses like
`UITable` which can just save the combined state and spread it to
`IrisGrid` since `IrisGrid` accepts both `GridState` and `IrisGridState`
props.

Added tests for `dehydrateIrisGridState`.

Improved the `memoize-one` type override to infer the types of the
equality function based on the actual function args.