feat: Add input filter support to GridWidgetPlugin (#2438)
This adds a few new hooks
- `useDhId` which will be used to route to a specific component in dh.ui
(via prop), or the panel ID (via context) as a fallback. This uses
`useFiber` which is not exported because it's an escape hatch that
should be used very rarely. `useDhId` takes no props and can read the
`__dhId` prop of the component via the `useFiber` hook.
- `useDashboardId` gives the ID of the current dashboard
- `useDashboardColumnFilters` is used to subscribe to column filter
changes and also advertise the columns and table (if provided) available
for filtering. This should allow us to hook up the existing
filtering/linking to dh.ui and probably also `ChartBuilder` in dh.ui
tables.

The contexts for `useDhId` and `useDashboardId` are added to `Dashboard`
so there shouldn't be any changes for enterprise since it uses the
`Dashboard` component.

Also reworked the events around this to have stronger types using the
helper methods we added.