fix: Grid making unnecessary onSelectionChanged calls (#2471)
Needed as part of DH-19292 to ensure we don't emit duplicate selection
events. Even with debouncing, we might emit the event, and then emit the
same selection (in a different array reference) on mouse up. This fixes
that so we don't emit the unnecessary events if the mouse up did not
remove selections

There are still some quirks around the selection events like when using
ctrl+click+drag to deselect part of a range, `onSelectionChanged` is
emitted with the subtracted portion as an extra range (e.g., `[[0, 5],
[1, 2]]`) until mouse up which commits the selection and removes the
overlapping ranges (`[[0, 0], [3, 5]]`). On the dh.ui side I'm just
consolidating the ranges, but would probably be good to fix that at some
point.