feat: DH-19722: Better handling for large numbers of open dashboards (#2481)
For DH-19722. This PR adds a popper for searching and filtering through
dashboards. It supports keyboard navigation and includes a global
shortcut (`Ctrl+Shift+D` / `Cmd+Shift+D`) to toggle the quick filter
popper.

Changes:
- Dashboard search: Filter dashboards by typing in the search input
- Global shortcut: Ctrl+Shift+D/Cmd+Shift+D opens the dashboard search
popper (only when 2+ dashboards exist)
- Keyboard interactions: arrow keys navigate between dashboard items;
enter sets the active dashboard to the currently focused item in the
list (if no dashboard in the list is focused it will default to the
first dashboard in the search)
- Adds 'endPlaceholder' to `SearchInput` to display the associated
shortcut