fix: DH-19601: Pin aggregations to the top in input tables (#2502)
For DH-19601. Input tables with aggregations in the footer were
experiencing conflicts between aggregated data and insertions to the
pending area. This pr pins aggregations to the top of input tables, that
way we have the bottom of the table reserved for pending rows, and we
don't need to worry about conflicts.