feat: Add URI support and api associated with widget to useWidget (#2509)
Part of DH-19001. Tested with my VM `mattrunyon-gplus` which has the
updated dh.ui and DHE web UI as well. Used the notebook w/ the ticket
number as several examples, and the `Demo` and `DemoDashboard` queries
w/ different users with different permissions to view those queries.
Tested using tables/figures/pickers and restarting the queries to ensure
they restarted and could continue being used.

This is mostly type changes to augment the object fetcher stuff to
accept a URI.

Also adds the associated `api` to the `useWidget` hook needed for PQ
URIs in DHE.

Renamed `useTableClose` to `useWidgetClose` as a more generic util and
deprecated `useTableClose`. They are the same, just more permissive
types since this can apply to non-tables