fix: IrisGridModelUpdater causes "cannot update during an existing state transition" warning (#2536)
`IrisGridModelUpdater` displays an error/warning in the browser console
when
- The reset button is clicked in the organize columns tab in the sidebar
of the grid
- A rollup is added to the table using the sidebar tab

These warnings were causing two e2e tests to fail locally. Full logs can
be found in the ticket.

Similar to https://github.com/deephaven/web-client-ui/pull/2249, it
looks like the `frozenColumns` setter in `IrisGridModelUpdater` was also
emitting an event in the render cycle that caused a React warning about
the render not being a pure function, because the event triggered a
state update in `IrisGrid`. The setter is now wrapped in `useEffect`
instead of `useOnChange` so the events are emitted after the render.