fix: DH-20229: Value is null or undefined when scrolling rollup (#2538)
The "Value is null or undefined" error was thrown from calling
`IrisGridTreeTableModel.textAlignForCell`when scrolling a table with
expanded rollups while hovering on a row.

I think these errors are due to `textAlignForCell` not being able to
find the row in the viewport when scrolling rapidly as it has already
left the viewport, and so it shouldn't matter what is returned in this
case if the cell isn't displayed on screen.
- In this case, the row is only used to determine whether the
constituent type should be used for formatting, and will now fallback to
the column type when the row is null.

To address potential issues with scrolling in the future, some tests
involving scrolling a table were added. These mainly represent the
failure scenario seen in the ticket, but more can be added in the future
to test other scenarios.
- From running the tests, a similar error was found in
`IrisGridRenderer.getExpandButtonPosition`, and a fallback value has
been added to replace the not null assertions there as well. This is the
same pattern used in
`IrisGridTextCellRenderer.ts.getCellOverflowButtonPosition`