refactor: DH-20098 React 19 support swap react-beautiful-dnd for @hello-pangea/dnd (#2548)
This should be all that is needed for DH-20098 after #2534

This replaces `react-beautiful-dnd` with `@hello-pangea/dnd` which is a
fork someone maintained after Atlassian stopped supporting
`react-beautiful-dnd`. It fixes the remaining React 18 warnings we see
in dev when rendering a `react-beautiful-dnd` list (default props not
allowed on functional component). The rollup/aggregations menu should
use these components as a test point.

I've published an alpha build of this branch and a customer said it was
working for them on React 19.